Allow two characters to trade items
a character to tries to give or take something will remember that they wanted this,
 so then if the other character does the complementary action it will succeed.

Example:
A >> inventory
A << You have a widget.

A >> give widget to B
A << B doesn't want the widget.

B >> inventory
B << A tries to give you a widget, but you don't take it.
B << You don't have anything.

B >> take widget
B << A gives you the widget.

B >> inventory
B << You have a widget.

A >> inventory
A << You give the widget to B.
A << You don't have anything.